bug fix on compilation for unix and win

+++ main.cpp
@@ -1,5 +1,6 @@
 #include "VoxWriter.h"
 
+
 // example from @unphased on this topic : https://github.com/aiekick/MagicaVoxel_File_Writer/issues/2
 
 int main() 
@@ -7,9 +8,8 @@ int main()
 	vox::VoxWriter vox;
 	for (int i = 0; i < 1000; ++i) {
 		for (int j = 0; j < 1000; ++j) {
-			vox.AddVoxel(i, j, (int)floor(sin((float)(i * i + j * j) / 50000) * 150) + 150, (i + j) % 255 + 1);
+			vox.AddVoxel(i, j, (int)std::floor(sinf((float)(i * i + j * j) / 50000) * 150) + 150, (i + j) % 255 + 1);
 		}
 	}
-
 	vox.SaveToFile("output_voxwriter.vox");
 }
